Built on the hello-triangle. It now moves
the hello-triangle now move by using a uniform variable in the vertex shader.
Also organized my util-lib library directory.
It contains rust document and rust document test to display how the function or datatype works with a test to prove it.
use cargo test to run the rust document test.